Install the Docusaurus package to `website`
All default files and comments were removed as they are not needed and
will be removed later as documentation and specific styling is added.

`docusaurus.config.js` was modified to enable "Docs-only mode" and
disable the blog. The footer, sidebar and navigation configuration were
adjusted to show only relevant information.