Create a `docs.yml` GitHub Action
This GitHub Action will set up a Node.js environment, checkout the
Git repository, install dependencies and run build scripts to generate
the Docusaurus website.

Only the `workflow_dispatch` event (manual trigger) is enabled to avoid
rebuilding the documentation too frequently. This may be changed to run
on a branch push as the repository reaches a stable state.

This GitHub Action is forked from the existing "Deploy to GitHub Pages"
Action, available on the GitHub Marketplace:
https://github.com/marketplace/actions/deploy-to-github-pages