Improve code coverage reporting for `packages/website-generator` to use source maps
Code coverage was accidentally set to check the coverage of the
compiled files. This meant any branches in logic introduced by the
compiler were being included.

By using source maps, only lines and branches in the source files are
reported against. This makes identifying uncovered lines and branches
easier as the reports reflect the source. Another benefit is coverage
will be more accurate because compiled code is no longer being included
in the reports.