Remove the `callback` argument from the `readFile` util
Handling callbacks was creating tricky type assertions so the callbacks
have been removed outside of the `readFile` method and placed alongside
where `readFile` is called.

Given that all of these instances are parsing text as JSON, a util
specifically for fetching a JSON file and handling errors could be a
good, more explicit, alternative.